requirements.txt
PyQt6>=6.4.0
requests>=2.28.0
pyobjc-core>=11.0  # macOS only
pyobjc-framework-Quartz>=11.0  # macOS only
